move forward declarations from Kokkos_View.hpp to a dedicated header
